Improve handling of keyboard, cursor and beeper
The cursor flash rate depends on the keyboard TX rate so make sure that
it's limited correctly. This means that beep updates will be generated
at the correct rate too; it appears that 48 are sent for a normal beep
so ignore that many repeats. Which gets exact beep recreation.

This showed that the 'alarm' generated timing wasn't sufficiently accurate
so that's been replaced by timing based on the vertical refresh emulation.
(If the CPU is frozen a simple 50ms sleep is used instead.)

With the good clock it is possible to generate a continous rollover
pattern when pasting keypresses to the terminal. Unfortunatly the VT100
ROM appears to be unable to decode this so the virtual keypresses are
NOT overlapped, just directly appended. This does give a significant
improvement in pasting speed.